remove unused function from innerSliderUtils.js
fix #1791
