+++ __tests__/sliderStyles.test.js
@@ -1,6 +1,5 @@
 import { mount } from "enzyme";
 import assign from "object-assign";
-import { getRequiredLazySlides } from "../src/utils/innerSliderUtils";
 import {
   createInnerSliderWrapper,
   clickNext,
